Update52

+++ scripts/osu.sh
@@ -11,65 +11,71 @@ drirc='
 '
 outputs=(DVI-I-1 VGA-1 VGA-0 eDP1 eDP-1 eDP-1-1 HDMI1 HDMI-1 HDMI-1-1)
 
-function osu {
+function changeres {
     if [ $USER == lukee ]; then
+        [[ $1 == 900 ]] && res="1440x900" || res="1920x1080"
         for output in "${outputs[@]}"; do
-            xrandr --output $output --mode 1440x900 --rate 74.98 2>/dev/null
-            [ $? == 1 ] && xrandr --output $output --mode 1400x900 2>/dev/null
+            xrandr --output $output --mode $res --rate 74.98 2>/dev/null
+            [ $? == 1 ] && xrandr --output $output --mode $res 2>/dev/null
         done
     fi
-    
+}
+
+function oss {
+    changeres 900
     if [[ $1 == "lazer" ]]; then
         $HOME/osu/osu.AppImage
     else
         echo "$drirc" > $HOME/.drirc
-        bash -c osukill
-        cd $HOME/osu # Or wherever you installed osu! in
+        wineserver -k
+        cd $HOME/osu
         wine osu!.exe "$@"
-        bash -c osukill
+        wineserver -k
         rm -f $HOME/.drirc
     fi
-    
-    if [ $USER == lukee ]; then
-        for output in "${outputs[@]}"; do
-            xrandr --output $output --mode 1920x1080 2>/dev/null
-            [ $? == 1 ] && xrandr --output $output --mode 1920x1080 2>/dev/null
-        done
-    fi
+    changeres
 }
 
 function random {
+    cd $HOME/osu
     for i in {1..4}; do
-        osu "$HOME/osu/oss/$(ls $HOME/osu/oss/ | shuf -n 1)"
+        wine osu!.exe "$HOME/osu/oss/$(ls $HOME/osu/oss/ | shuf -n 1)"
     done
 }
 
 function remove {
-    osslist=$(ls $HOME/osu/oss/ | sed -e 's/\.osz$//')
-    osulist=$(ls $HOME/osu/Songs)
-    ossremoved=$(comm -12 $osslist $osulist)
-    sed -i 's/$/.osz/' $ossremoved
-    sed -i 's/^/oss\//' $ossremoved
+    ls $HOME/osu/oss/ | sed -e 's/\.osz$//' | tee osslist
+    ls $HOME/osu/Songs | tee osulist
+    ossremoved=$(comm -12 osslist osulist)
+    sed 's/$/.osz/' $ossremoved
+    sed 's/^/oss\//' $ossremoved
     cat $ossremoved | xargs -d '\n' rm -rf
+    rm -f osslist osulist
     cat $ossremoved
 }
 
 function update {
     cd $HOME/osu
+    osuapi=$(curl -s https://api.github.com/repos/ppy/osu/releases/latest)
     current=$(cat osu.AppImage.version 2>/dev/null)
-    latest=$(curl -s https://api.github.com/repos/ppy/osu/releases/latest | grep "tag_name" | cut -d : -f 2,3)
+    [ ! $current ] && current='N/A'
+    latest=$(echo "$osuapi" | grep "tag_name" | cut -d : -f 2,3)
+    echo "osu!lazer"
+    echo "* Current version: $current"
+    echo "* Latest version: $latest"
     if [[ $latest != $current ]]; then
         rm osu.AppImage* 2>/dev/null
-        curl -s https://api.github.com/repos/ppy/osu/releases/latest | grep "/osu.AppImage" | cut -d : -f 2,3 | tr -d \" | wget -nv --show-progress -i -
+        echo "$osuapi" | grep "/osu.AppImage" | cut -d : -f 2,3 | tr -d \" | wget -nv --show-progress -i -
+        [ ! -e osu.AppImage ] && echo "Update failed" && exit
         chmod +x osu.AppImage
         echo "$latest" > osu.AppImage.version
-        echo "Updated"
+        echo "Updated osu!lazer to $latest"
     else
         echo "Currently updated, nothing to do"
     fi
 }
 
-function osuinstall {
+function install {
     sudo cp /etc/security/limits.conf /etc/security/limits.conf.bak
     echo "@audio - nice -20
     @audio - rtprio 99" | sudo tee /etc/security/limits.conf
@@ -90,7 +96,7 @@ function osuinstall {
     default-fragments = 2
     default-fragment-size-msec = 4" | sudo tee /etc/pulse/daemon.conf.d/10-better-latency.conf
 
-    sudo cp $(dirname $(type -p $0))/osu.sh /usr/bin/osu
+    [ ! -e /usr/bin/osu ] && sudo cp $(dirname $(type -p $0))/osu.sh /usr/bin/osu
     sudo chmod +x /usr/bin/osu
 
     mkdir $HOME/.config/pulse 2>/dev/null
@@ -98,6 +104,12 @@ function osuinstall {
     sed -i "s/load-module module-udev-detect.*/load-module module-udev-detect tsched=0 fixed_latency_range=yes/" $HOME/.config/pulse/default.pa
     
     sudo pacman -S --noconfirm --needed lib32-alsa-plugins lib32-gnutls lib32-libxcomposite winetricks
+    if [ -d $HOME/.wine_osu ]; then
+        read -p "wine_osu folder detected! Delete and reinstall? (y/N) " Confirm
+        if [[ $osuDL == y ]] || [[ $osuDL == Y ]]; then
+            exit
+        fi
+    fi
     rm -rf $HOME/.wine_osu
     
     winetricks dotnet40
@@ -120,9 +132,10 @@ elif [[ $1 == "remove" ]]; then
 elif [[ $1 == "update" ]]; then
     update
 elif [[ $1 == "lazer" ]]; then
-    osu lazer
+    oss lazer
 elif [[ $1 == "kill" ]]; then
     wineserver -k
+    changeres
     exit
 elif [[ $1 == "help" ]]; then
     echo "Usage: $0 <operation> [...]"
@@ -135,7 +148,7 @@ elif [[ $1 == "help" ]]; then
     osu {remove}
     osu {update}"
 elif [[ $1 == "install" ]]; then
-    osuinstall
+    install
 else
-    osu "$@"
+    oss "$@"
 fi
